+++ src/Components/Config/ConfigApi.php
@@ -7,14 +7,16 @@ namespace InnStudio\Prober\Components\Config;
 
 class ConfigApi
 {
-    public static $APP_VERSION                 = '3.8';
-    public static $APP_NAME                    = 'X Prober';
-    public static $APP_URL                     = 'https://github.com/kmvan/x-prober';
+    public static $APP_VERSION                  = '3.9';
+    public static $APP_NAME                     = 'X Prober';
+    public static $APP_URL                      = 'https://github.com/kmvan/x-prober';
     public static $APP_CONFIG_URLS              = array('https://raw.githubusercontent.com/kmvan/x-prober/master/AppConfig.json', 'https://api.inn-studio.com/download/?id=xprober-config');
-    public static $APP_CONFIG_URL_DEV          = 'http://127.0.0.1/prober/AppConfig.json';
-    public static $AUTHOR_URL                  = 'https://inn-studio.com/prober';
-    public static $UPDATE_PHP_URLS             = array('https://raw.githubusercontent.com/kmvan/x-prober/master/dist/prober.php', 'https://api.inn-studio.com/download/?id=xprober');
-    public static $AUTHOR_NAME                 = 'INN STUDIO';
-    public static $LATEST_PHP_STABLE_VERSION   = '7';
-    public static $LATEST_NGINX_STABLE_VERSION = '1.16.1';
+    public static $APP_CONFIG_URL_DEV           = 'http://127.0.0.1/prober/AppConfig.json';
+    public static $APP_TEMPERATURE_SENSOR_URL   = 'http://127.0.0.1';
+    public static $APP_TEMPERATURE_SENSOR_PORTS = array(2048, 4096);
+    public static $AUTHOR_URL                   = 'https://inn-studio.com/prober';
+    public static $UPDATE_PHP_URLS              = array('https://raw.githubusercontent.com/kmvan/x-prober/master/dist/prober.php', 'https://api.inn-studio.com/download/?id=xprober');
+    public static $AUTHOR_NAME                  = 'INN STUDIO';
+    public static $LATEST_PHP_STABLE_VERSION    = '7';
+    public static $LATEST_NGINX_STABLE_VERSION  = '1.16.1';
 }